requirements.txt
alembic~=1.13.2
asttokens~=2.4.1
certifi~=2024.2.2
charset-normalizer~=3.3.2
click~=8.1.7
colorama~=0.4.6
colorlog~=6.8.2
comm~=0.2.2
debugpy~=1.8.1
decorator~=5.1.1
exceptiongroup~=1.2.1
executing~=2.0.1
greenlet~=3.0.3
idna~=3.7
importlib_metadata~=7.1.0
ipykernel~=6.29.4
ipython
jedi~=0.19.1
joblib~=1.4.0
jupyter_client~=8.6.1
jupyter_core~=5.7.2
Mako~=1.3.5
MarkupSafe~=2.1.5
matplotlib-inline~=0.1.7
nest-asyncio~=1.6.0
nltk~=3.8.1
numpy
optuna~=3.6.1
packaging~=24.0
pandas
parso~=0.8.4
platformdirs~=4.2.1
prompt-toolkit~=3.0.43
psutil~=5.9.8
pure-eval~=0.2.2
Pygments~=2.17.2
python-dateutil~=2.9.0.post0
pytz~=2024.1
PyYAML~=6.0.1
pyzmq~=26.0.2
regex~=2024.4.16
requests~=2.31.0
scikit-learn
scipy
six~=1.16.0
SQLAlchemy~=2.0.31
stack-data~=0.6.3
threadpoolctl~=3.5.0
tornado~=6.4
tqdm~=4.66.2
traitlets~=5.14.3
typing_extensions~=4.11.0
tzdata~=2024.1
urllib3~=2.2.1
wcwidth~=0.2.13
zipp~=3.18.1
